Refactor UI components, formatting and small fixes
Apply widespread formatting and readability improvements across Razor and C# files (multiline ternaries, attribute wrapping, spacing). Reorganize and introduce several fields and constants (dialog options, dialog state fields like title/author/notes, cached/selection state, example scripts, property records) and move/dispose logic for refresh handlers. Update SaveFileInfo, SaveFileRepair, ShowdownImport, BatchEditor, Stats, Legality and bank/teams/party UI for clearer layout and minor behavior fixes (e.g. legalization/status handling, parsing call update, snackbar messages, memory stream usage). Overall changes are intended to improve code clarity and correct small issues without major feature additions.

+++ Pkmds.Tests/PokemonSlotLegalityTests.cs
@@ -1,4 +1,3 @@
-using Bunit;
 using Microsoft.AspNetCore.Components;
 
 namespace Pkmds.Tests;
@@ -15,7 +14,7 @@ public class PokemonSlotLegalityTests
         using var ctx = BunitTestHelpers.CreateBunitContext(appState, refreshService, appService);
 
         var pkm = saveFile.GetPartySlotAtIndex(0);
-        pkm.Species.Should().BeGreaterThan((ushort)0, "need a real Pokémon for the legality overlay to appear");
+        pkm.Species.Should().BeGreaterThan(0, "need a real Pokémon for the legality overlay to appear");
 
         var cut = ctx.Render<PokemonSlotComponent>(p => p
             .Add(c => c.SlotNumber, 0)
